Simplified code: removed the custom field for zoom
