+++ web/client/themes/default/less/annotations.less
@@ -118,20 +118,6 @@
 .mapstore-annotations-info-viewer-items {
     height: ~'calc(100% - 50px)';
     overflow: auto;
-
-    .mapstore-annotations-info-viewer-zoom {
-        position: absolute;
-        top: 60px;
-        left: 192px;
-    }
-}
-
-#mapstore-annotations-panel {
-    .mapstore-annotations-info-viewer-zoom {
-        position: absolute;
-        top: 49px;
-        left: 299px;
-    }
 }
 
 .mapstore-right-panel .react-swipeable-view-container > div {
